requirements.txt
torch==1.3.1
torchvision==0.4.2
numpy==1.17.4
opencv-python
packaging
cython
pycocotools
matplotlib
tqdm
mmcv==0.2.15
pillow==6.1
albumentations
imagecorruptions
chainer